Fixed issues causing travis build failures
